requirements.txt
streamlit>=1.50.0
python-dotenv>=0.19.0
